requirements.txt
flask
requests
python-binance
ta
pandas
gunicorn